feat/useAuth함수 타입 정의
